+++ packages/components/organisms/f-checkout/src/store/checkout.module.js
@@ -23,6 +23,10 @@ import {
 
 import checkoutIssues from '../checkout-issues';
 
+/**
+ * @param {String} code - The code returned by an API.
+ * @returns {object} - An object with the issue's desired behaviours and the code.
+ */
 const getIssueByCode = code => {
     const issue = checkoutIssues[code];
 
@@ -31,6 +35,32 @@ const getIssueByCode = code => {
     }
 
     return null;
+};
+
+/**
+* @function resolveCustomerDetails
+* If certain customer details are missing from api `data.customer` object then fallback
+* to the decoded `state.AuthToken` details and re-assign back to `data.customer`.
+* @param  {object} data  - Api response object.
+* @param  {object} state - The current `checkout` state.
+*/
+const resolveCustomerDetails = (data, state) => {
+    if (data && data.customer) {
+        let tokenData;
+
+        if (!data.customer.phoneNumber) {
+            tokenData = jwtDecode(state.authToken);
+
+            data.customer.phoneNumber = tokenData.mobile_number || tokenData.phone_number;
+        }
+
+        if (!data.customer.firstName || !data.customer.lastName) {
+            tokenData = tokenData || jwtDecode(state.authToken);
+
+            data.customer.firstName = tokenData.given_name;
+            data.customer.lastName = tokenData.family_name;
+        }
+    }
 };
 
 export default {
@@ -103,16 +133,7 @@ export default {
 
             const { data } = await axios.get(url, config);
 
-            /**
-            * If no `customer.phoneNumber` from api then read from decoded
-            * AuthToken and re-assign mobile number or phone number
-            * to `data.customer.phoneNumber`
-            */
-            if (data && data.customer && !data.customer.phoneNumber) {
-                const tokenData = jwtDecode(state.authToken);
-
-                data.customer.phoneNumber = tokenData.mobile_number || tokenData.phone_number;
-            }
+            resolveCustomerDetails(data, state);
 
             commit(UPDATE_STATE, data);
 
@@ -268,23 +289,6 @@ export default {
         },
 
         /**
-         * Get the customer name from JWT claims and update state with the result
-         *
-         * @param {Object} context - Vuex context object, this is the standard first parameter for actions
-         * @param {Object} payload - Parameter with the different configurations for the request.
-         */
-        getCustomerName: async ({ commit, state }) => {
-            const tokenData = jwtDecode(state.authToken);
-
-            const customer = {
-                firstName: tokenData.given_name,
-                lastName: tokenData.family_name
-            };
-
-            commit(UPDATE_CUSTOMER_DETAILS, customer);
-        },
-
-        /**
          * Post the order details to the Order Placement API and get the `orderId` from the response.
          *
          * @param {Object} context - Vuex context object, this is the standard first parameter for actions
@@ -401,6 +405,7 @@ export default {
 
             if (customer) {
                 state.customer.firstName = customer.firstName;
+                state.customer.lastName = customer.lastName;
                 state.customer.mobileNumber = customer.phoneNumber;
             }
 
@@ -429,7 +434,7 @@ export default {
 
         [UPDATE_AUTH_GUEST]: (state, authToken) => {
             state.authToken = authToken;
-            state.isLoggedIn = false;
+            state.isLoggedIn = true;
             state.isGuestCreated = true;
         },